requirements.txt
flask
flask-cors
openai-whisper
sounddevice
numpy
scipy
pydub